fix: Handle payments canceled by Stripe
An error happened to someone who initiated a payment few days ago but
didn't finish the process. The payment was considered as "waiting" by
Flus, but Stripe canceled it. When the user was checking the status of
the payment on Stripe, the later was telling him the session expired.
Back to Flus, we were telling Stripe to cancel it (again), which raised
an error.

This is an quick and easy fix, but a better one would be to setup a
webhook endpoint to receive updates from Stripe.

References:

- https://github.com/flusio/xExtension-Flus/blob/master/lib/stripe-php/lib/PaymentIntent.php#L69-L83
- https://stripe.com/docs/webhooks